feat: introduce Blockaid transaction scanning configuration and implementation, enhancing security checks
