+++ core/chain/security/blockaid/config.ts
@@ -1,0 +1,4 @@
+import { rootApiUrl } from '@core/config'
+
+export const blockaidBaseUrl = `${rootApiUrl}/blockaid/v0`
+export const blockaidClientId = 'vultisig-desktop-extension'
